Move consts to covered_files
