+++ src/reuse/__init__.py
@@ -22,7 +22,6 @@ change the public API.
 import gettext
 import logging
 import os
-import re
 from dataclasses import dataclass, field
 from enum import Enum
 from importlib.metadata import PackageNotFoundError, version
@@ -52,43 +51,6 @@ if gettext.find("reuse", localedir=_LOCALE_DIR):
     _LOGGER.debug("translations found at %s", _LOCALE_DIR)
 else:
     _LOGGER.debug("no translations found at %s", _LOCALE_DIR)
-
-
-_IGNORE_DIR_PATTERNS = [
-    re.compile(r"^\.git$"),
-    re.compile(r"^\.hg$"),
-    re.compile(r"^\.sl$"),  # Used by Sapling SCM
-    re.compile(r"^LICENSES$"),
-    re.compile(r"^\.reuse$"),
-]
-
-_IGNORE_MESON_PARENT_DIR_PATTERNS = [
-    re.compile(r"^subprojects$"),
-]
-
-_IGNORE_FILE_PATTERNS = [
-    # LICENSE, LICENSE-MIT, LICENSE.txt
-    re.compile(r"^LICEN[CS]E([-\.].*)?$"),
-    re.compile(r"^COPYING([-\.].*)?$"),
-    # ".git" as file happens in submodules
-    re.compile(r"^\.git$"),
-    re.compile(r"^\.hgtags$"),
-    re.compile(r".*\.license$"),
-    re.compile(r"^REUSE\.toml$"),
-    # Workaround for https://github.com/fsfe/reuse-tool/issues/229
-    re.compile(r"^CAL-1.0(-Combined-Work-Exception)?(\..+)?$"),
-    re.compile(r"^SHL-2.1(\..+)?$"),
-]
-
-_IGNORE_SPDX_PATTERNS = [
-    # SPDX files from
-    # https://spdx.github.io/spdx-spec/conformance/#44-standard-data-format-requirements
-    re.compile(r".*\.spdx$"),
-    re.compile(r".*\.spdx.(rdf|json|xml|ya?ml)$"),
-]
-
-# Combine SPDX patterns into file patterns to ease default ignore usage
-_IGNORE_FILE_PATTERNS.extend(_IGNORE_SPDX_PATTERNS)
 
 
 class SourceType(Enum):
